Moved door enums to FaceEnums.h
